Extract cache logic from providers
Instead of a collection of providers that also need to implement cache
logic, we now have providers that only task is to provide CLDR data,
and a variety of cache to cache that data. CachedProvider integrates
a provider and a cache, fetching data from the provider and updating
the cache when necessary. CacheCollection uses multiple cache
strategies to retrieve and store data.